#4
Chuẩn hóa response

+++ src/main/java/vn/hoidanit/jobhunter/controller/UserController.java
@@ -4,19 +4,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 import vn.hoidanit.jobhunter.domain.User;
 import vn.hoidanit.jobhunter.service.UserService;
+import vn.hoidanit.jobhunter.service.error.IdInvalidException;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.PutMapping;
-
-
-
 
 @RestController
 public class UserController {
@@ -27,22 +26,29 @@ public class UserController {
         this.userService = userService;
     }
 
-    @PostMapping("/user")
-    public User createNewUser(@RequestBody User RqUser){
-        User newUser= this.userService.setNewUser(RqUser);
-        return newUser;
+    @PostMapping("/users")
+    public ResponseEntity<User> createNewUser(@RequestBody User RqUser) {
+        User newUser = this.userService.setNewUser(RqUser);
+        // return newUser;
+        return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
     }
 
-    @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable long id){
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable long id) throws IdInvalidException {
+      
+        if(id >1111){
+            throw new IdInvalidException("ID must be less than 1112");
+        }
+        // hoặc viết kiểu ngắn hơn
         this.userService.deleteUserById(id);
-        return "user deleted";
+        // return ResponseEntity.status(HttpStatus.OK).body("Deleted");
+        return ResponseEntity.ok("Deleted");
     }
 
-    @GetMapping("/user/{id}")
-    public User getUserById(@PathVariable long id) {
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> getUserById(@PathVariable long id) {
         User user = this.userService.getUserById(id);
-        return user;
+        return ResponseEntity.ok(user);
     }
 
     @GetMapping("/users")
@@ -51,7 +57,7 @@ public class UserController {
         return users;
     }
 
-    @PutMapping("user")
+    @PutMapping("/users")
     public User updateUser(@RequestBody User updatedUser) {
         User user = this.userService.setNewUser(updatedUser);
         return user;
